+++ tests/features/stdin_input_steps.py
@@ -16,13 +16,14 @@ def i_create_all_resources_to_test_from_stdin(step, data=None, test=None, output
         assert False
     command = ("cat " + test + "|bigmler --train " + data +
                " --test --store --output " + output + " --max-batch-models 1")
-    shell_execute(command, output, test=None)
+    shell_execute(command, output, test=test)
 
 
 @step(r'I create a BigML source from stdin using train "(.*)" file and logging in "(.*)"$')
-def i_create_source_from_stdin(step, data=None, output=None):
-    if data is None or output is None:
+def i_create_source_from_stdin(step, data=None, output_dir=None):
+    if data is None or output_dir is None:
         assert False
     command = ("cat " + data + "|bigmler --train " +
-               "--store --no-dataset --no-model --output " + output + " --max-batch-models 1")
-    shell_execute(command, output, test=None)
+               "--store --no-dataset --no-model --output-dir " +
+               output_dir + " --max-batch-models 1")
+    shell_execute(command, output_dir + "/test", test=None)